doc-verify-automation-updated/requirements.txt
pandas
pypdf
pytesseract
opencv-python
Pillow
rapidfuzz
numpy
regex
openpyxl
